mmngr_drv: mmngr: return -ENOMEM if kzalloc() fails
Currently on failure the value returned from kzalloc() is -1.
It should return -ENOMEM.

Signed-off-by: Veeraiyan Chidambaram <veeraiyan.chidambaram@in.bosch.com>

The original patch also update return at
...
		/* Allocate bitmap for entry */
 		bm_lossy = kzalloc(sizeof(struct BM), GFP_KERNEL);
-		if (bm_lossy == NULL)
+		if (bm_lossy == NULL) {
+			ret = -ENOMEM;
 			break;
...
but due to the design of the fallback mode in Lossy,
the above update corrupted it.

Another patch was introduced later by the same author to fix it,
so merged the two patches for convenience.

------ Fix patch description ------
During mmngr_drv driver clean up init_lossy_info function
has been modified to return negative error code for kzalloc failure.
When above change was submitted to Renesas via mantis ticket [1],
it was rejected as this goes against the design of fallback mode in Lossy.

[1] https://ass.renesas.eu/mantis/view.php?id=7744#c56561

Fixes: commit ("mmngr_drv: mmngr: return -ENOMEM if kzalloc() fails")
-----------------------------------

Signed-off-by: Nam Nguyen <nam.nguyen.yh@renesas.com>