mmngr_drv: mmngr: check result when setting DMA mode
The `dma_set_mask_and_coherent` does not always perform successfully
and it also might fail. Ignoring this error and attempting to use DMA
in case of failure will lead to undefined behavior.
See for details: Documentation/core-api/dma-api-howto.rst

Also, this patch puts the second call of the function inside the
conditional block based on `IPMMU_MMU_SUPPORT` value. This is necessary
to avoid a double function call when the `IPMMU_MMU_SUPPORT` is true.

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>

Update return error from '-1' to 'ret'.

Signed-off-by: Nam Nguyen <nam.nguyen.yh@renesas.com>