mmngr_drv: mmngr: limit min size for MM_KERNELHEAP
When the MMNGR requested a memory via the `mm_ioc_alloc` (in other words
with `MM_KERNELHEAP` flag) it tries to allocate a memory from the common
CMA area. MMNGR sets a NULL to the device CMA information and then
request memory uses DMA API. Kernel sees that device doesn't have it's
own CMA area and tries to allocate memory from the common CMA area.

In most cases it works as expected, but in some cases Kernel may returns
memory outside of the common CMA area:
 - when CMA memory runs out;
 - when requested memory size less or equal than one memory page
   (because a single page is always continuous by itself and can be
   allocated elsewhere).
In such cases MMNGR gets memory outside a common CMA area. Moreover the
allocated memory may not match any item of the MMNGR transformation
table! Eventually this leads to problems.

The current change is addressed to cases when the requested size less
then a memory page and Kernel returns a usual DMA addressable memory
(but not CMA one) to save an expencive CMA memory and prevent it's
fragmentation.

An alternative solution would be allocation using directly CMA API.
Unfortunately it is not possible in this particular case, because memory
may be allocated only from the own CMA area. But the MMNGR doesn't own a
common CMA area (in fact a common CMA area is owned by Contiguous Memory
Allocator for DMA mapping framework). So using CMA API to allocate a
memory from the common CMA area isn't suitable.

See for details:
 - include/linux/dma-mapping.h;
 - kernel/dma/mapping.c;
 - kernel/dma/direct.c;
 - kernel/dma/contiguous.c;
 - mm/cma.c.

The changes are applied only for configuration with enabled IPMMU
because if IPMMU is disabled using memory outside a CMA region will not
lead to problems.

The solution was tested on Kernel v5.10.41.

Test: request a memory via libmmngr (MMNGR userspace library) with a
      `MMNGR_VA_SUPPORT` flag and size less then `PAGE_SIZE`; check the
      returned address and compare it with range of a common CMA area.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>