Adapting patch "From c860f4a10a2ba42de7be053410163a65bc4ca642 Mon Sep 17 00:00:00 2001 From: Sergii Piatakov <sergii.piatakov@globallogic.com> Date: Wed, 25 Jan 2023 09:31:19 +0000 Subject: [PATCH] mmngr_drv: mmngr: check result when setting DMA mode
The `dma_set_mask_and_coherent` does not always perform successfully
and it also might fail. Ignoring this error and attempting to use DMA
in case of failure will lead to undefined behavior.
See for details: Documentation/core-api/dma-api-howto.rst

Also, this patch puts the second call of the function inside the
conditional block based on `IPMMU_MMU_SUPPORT` value. This is necessary
to avoid a double function call when the `IPMMU_MMU_SUPPORT` is true.

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>"