Adapting patch "From f6e775c96159f979153966b144100820c20a23cb Mon Sep 17 00:00:00 2001 From: Oleksii Khramkov <oleksii.khramkov@globallogic.com> Date: Sun, 6 Aug 2023 21:21:18 +0300 Subject: [PATCH] mmngr_drv: mmngr: set 32-bit DMA mask for M3N and E3
With the current implementation, the MMNGR sets a 40-bit DMA bitmask
(when IPMMU is enabled) regardless of the actual SoC. At the same
time, M3N and E3 don't have 40-bit address space. IPMMU MMU supports
physical address space in the legacy area.
It is proposed to make changes that allow using the correct DMA
bitmask for all devices.

Signed-off-by: Oleksii Khramkov <oleksii.khramkov@globallogic.com>
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>"